after the player has dismissed the alert popup.
Such a block of code is called a closure.
{action in self.startNewRound(); self.updateLabels()}

This is also known as the callback pattern.

Normally you don’t need to use self to send messages to the view controller, even though it is allowed. The exception: inside closures you do have to use self to refer to the view controller.